pyproject.toml
[project]
name = "mcp-server"
version = "0.2.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.8.1",
    "httpx>=0.28.1",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.34.3",
]
